test: refactor into multiple files
